http options float numbers

+++ src/Providers/HttpClientOptions.php
@@ -10,8 +10,8 @@ class HttpClientOptions
      * @param array<string, string|int|float>|null $headers
      */
     public function __construct(
-        public readonly ?int $timeout = null,
-        public readonly ?int $connectTimeout = null,
+        public readonly ?float $timeout = null,
+        public readonly ?float $connectTimeout = null,
         public readonly ?array $headers = null,
     ) {
     }
